backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fish-audio-sdk[utils]>=1.1.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "requests>=2.32.5",
]
